% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/along.R
\name{along_rows}
\alias{along_rows}
\alias{along_cols}
\title{Along a table dimension}
\usage{
along_rows(x)

along_cols(x)
}
\arguments{
\item{x}{a table like structure.}
}
\description{
functions as \code{\link{seq_along}} but for rows or columns
of a table like structure.
}
