% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\name{cd}
\alias{cd}
\title{Change Directory}
\usage{
cd(new = NULL)
}
\arguments{
\item{new}{new working directory}
}
\value{
Always returns the current working directory.
}
\description{
Change the current working directory or get the current working directory
will also follow windows link files.
}
