% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Confidence interval structure}
\usage{
ci(estimate, lower, upper, confidence = 0.95, ...)
}
\arguments{
\item{estimate}{The Estimate}

\item{lower}{Lower bound}

\item{upper}{Upper bound.}

\item{confidence}{confidence level}

\item{...}{other information such as}
}
\description{
Create a \code{confidence-interval} object.
}
