% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{inline}
\alias{inline}
\alias{\%\n\%}
\alias{\%<<\%}
\alias{\%<<<\%}
\alias{\%~\%}
\alias{\%!~\%}
\alias{\%or\%}
\alias{\%inherits\%}
\alias{\%is a\%}
\title{Inline shortcuts and aliases}
\usage{
lhs \%\\n\% rhs

lhs \%<<\% rhs

lhs \%<<<\% rhs

lhs \%~\% rhs

lhs \%!~\% rhs

lhs \%or\% rhs

lhs \%inherits\% rhs

lhs \%is a\% rhs
}
\arguments{
\item{lhs}{left hand side}

\item{rhs}{right hand side}
}
\description{
A collection of inline operators to simplify and make your code
prettier.
}
\section{Functions}{
\itemize{
\item \code{\%\\n\%}: Concatenate with new line.

\item \code{\%<<\%}: Concatenate with space

\item \code{\%<<<\%}: Concatenate without space

\item \code{\%~\%}: inline pattern match

\item \code{\%!~\%}: inline pattern "does not match"

\item \code{\%or\%}: Return x unless \code{length(lhs)==0} then return \code{rhs}.

\item \code{\%inherits\%}: Inline alias of \code{\link[base]{inherits}}

\item \code{\%is a\%}: Alternative form of inherits.
}}

\examples{
"Hello" \%\\n\% "world"
"Hello" \%<<\% "world"
"Hello" \%<<<\% "world"
.T(Hello, world)  \%~\% "world"
character(0) \%or\% "pick me"
NULL \%or\% "no pick me"
"Hello world" \%inherits\% "character"


}
