% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mimetype}
\alias{mimetype}
\title{Mimetype}
\usage{
mimetype(file)
}
\arguments{
\item{file}{file to get information for.}
}
\description{
Infer the mime type of a file, if on windows will attempt to 
read the mime type info from the registry.
If a matching type cannot be found the generic
"application/octet-stream" will be returned.
}
