% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{Min}
\alias{Min}
\alias{Max}
\alias{Mean}
\alias{Median}
\title{Convenience Wrappers}
\usage{
Min(..., na.rm = TRUE)

Max(..., na.rm = TRUE)

Mean(..., na.rm = TRUE)

Median(..., na.rm = TRUE)
}
\arguments{
\item{...}{Other Arguments}

\item{na.rm}{Remove missing?}
}
\description{
These functions are for convenience and modify the original
function to exclude missing by default.
These are also highly useful to avoid warnings when converting to
SQL when using \link{dbplyr} functions and remote tables.
}
