% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_stars.R
\name{p_stars}
\alias{p_stars}
\title{P-value significance symbols}
\usage{
p_stars(p.value)
}
\arguments{
\item{p.value}{p-value to generate symbols for.}
}
\description{
Create significance symbols for p-values.
}
\examples{
p_stars(c(0.5, 0.09, 0.04, 0.009, 0.0009, 0))

}
