% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\alias{as.character.percent}
\alias{format.percent}
\title{Designate a numeric vector as a percent.}
\usage{
percent(x, ...)

\method{as.character}{percent}(x, places = NULL, threshold = NULL, ...)

\method{format}{percent}(x, places = NULL, threshold = NULL, ...)
}
\arguments{
\item{x}{a numeric object.}

\item{...}{additional formatting arguments.}

\item{places}{Number of places after decimal to show.}

\item{threshold}{Minimum percent to show.}
}
\description{
Designate a numeric vector as a percent.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.character}: Convert to character.

\item \code{format}: format the percent.
}}

