% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{A "Lazy" read version}
\usage{
read(file, ..., verbose = FALSE, glob = TRUE)
}
\arguments{
\item{file}{The file(s) or pattern of files to read.}

\item{...}{options to pass to the reader functions.}

\item{verbose}{Print messages?}

\item{glob}{interpret patterns as globs?}
}
\description{
We often find that we are reading the same type of files over and
over again.  Read is a lazy version of the read.* functions.
It tries to intelligently pick the appropriate reader from
file extension or mime type info.
}
\details{
read will first check if the file exists as specified.
If the file is not found then it will check if the file
represents a pattern converted to a regular expression
through \code{\link[utils]{glob2rx}} if glob=TRUE (default).
Note that only one pattern can be considered if file is
of length greater than 1 it will assume that exact file
names are given.

If multiple files are given or found through pattern
match then the results are given in a list, otherwise
a single object is returned. For elements
of said list, as in the case of a single result,
an attempt will be made to convert results to
a \link{tibble} through \code{\link[tibble]{as_tibble}}.

Read will first look for readers
with the following priority
\enumerate{
\item Known readers with the given name
\item Known readers registered to handle the given mime-type
\item Known readers registered to handle the given file extension
\item Any loaded function of the form \code{read.<<ext>>}
\item Any loaded function of the form \code{read_<<ext>>}
\item Any loaded function of the form \code{read<<ext>>}
}
where \code{<<ext>>} is replaced by the extension of the file.
}
\examples{
\dontrun{
# reads all the xlsx files in a directory
read("*.xlsx")

# reads all the xlsx files in the data directory
read("data/*.xlsx")


}

}
