% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_release.R
\name{read_description}
\alias{read_description}
\title{Read a package DESCRIPTION file}
\usage{
read_description(file, ..., text = NULL)
}
\arguments{
\item{file}{the path to the "DESCRIPTION" file.}

\item{...}{passed to \code{\link[=read.dcf]{read.dcf()}}.}

\item{text}{optional to specify content as a character vector of lines.}
}
\description{
Read a package "DESCRIPTION" file, returning
the results as a \code{\link[=tibble]{tibble()}} with the appropriate fields
parsed and evaluated. List fields, depends, imports, suggests
and enhances, are converted into nested lists.
}
