% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sift.R
\name{sift}
\alias{sift}
\alias{sieve}
\title{Filter a character vector}
\usage{
sift(x, pattern, value = TRUE, perl = TRUE, ignore.case = TRUE,
  on.names = FALSE, invert = FALSE, ...)

sieve(..., invert = TRUE)
}
\arguments{
\item{x}{a character vector where matches are sought, or an
    object which can be coerced by \code{as.character} to a character
    vector.  \link{Long vectors} are supported.}

\item{pattern}{character string containing a \link{regular expression}
    (or character string for \code{fixed = TRUE}) to be matched
    in the given character vector.  Coerced by
    \code{\link{as.character}} to a character string if possible.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  Missing values are allowed except for
    \code{regexpr} and \code{gregexpr}.}

\item{value}{if \code{FALSE}, a vector containing the (\code{integer})
    indices of the matches determined by \code{grep} is returned, and if
    \code{TRUE}, a vector containing the matching elements themselves is
    returned.}

\item{perl}{logical.  Should Perl-compatible regexps be used?}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{on.names}{Perform filtering on the names of the vector instead.}

\item{invert}{logical.  If \code{TRUE} return indices or values for
    elements that do \emph{not} match.}

\item{...}{Passed on}
}
\description{
\code{sift} filters the list to elements matching the pattern.
\code{sieve} filters them out.
}
