% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T.R
\name{.T}
\alias{.T}
\title{Lazily create a character vector}
\usage{
.T(...)
}
\arguments{
\item{...}{objects to quote}
}
\description{
Lazily create a character vector.
Any argument passed to .T will be converted to
a string without evaluating the symbols.
}
\examples{
.T(hell0, world)
.T(1+2)
.T("this how you include spaces")

}
