% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1.R
\name{table_1}
\alias{table_1}
\alias{table_1_dispatcher}
\alias{table_1_summarise}
\title{Create a Table 1}
\usage{
table_1(.data, key, ...)

table_1_dispatcher(.data, var, name, key)

table_1_summarise(.data, var, name, key)
}
\arguments{
\item{.data}{a dataset}

\item{key}{the comparison variable, such as case/control.}

\item{...}{a lazy list of variables to include in the description.}

\item{var}{Variable identifier, used to dispatch}

\item{name}{name of the variable}
}
\description{
This helps creates the demographics table, 
the eponymous "table 1".
Given a data set a key for columns, 
describe the differences across the provided 
factor variables between the levels of key.
}
\details{
`table_1_summarise` and `table_1_dispatcher` dispatch on the data type of the 
variable identified by `var` in `.data`
}
