% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/using.R
\name{using}
\alias{using}
\title{Intelligently load libraries}
\usage{
using(..., install.missing = getOption("using::install.missing",
  !is.null(getOption("repos"))), quiet = getOption("using::quiet",
  interactive()))
}
\arguments{
\item{...}{a list of packages to install, evaluated lazily.}

\item{install.missing}{Should an attempt be made to install 
packages if they are not installed?}

\item{quiet}{run without messages.}
}
\description{
Load a list of libraries.  
If the library is not installed install it and load it.
If using is called durring R startup, such as in a
.Rprofile file append the package(s) to the defaultPackages 
option to induce loading at the appropriate time.
}
\examples{
\dontrun{
using(tools)
}
}
