% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{with_margins}
\alias{with_margins}
\title{Operate over margins}
\usage{
with_margins(FUN, all.name = "(All)")
}
\arguments{
\item{FUN}{the function to wrap.}

\item{all.name}{the string to use to represent that a variable was marginalized over.}
}
\description{
Wrap a function to operate over groups and all possible
marginals of groupings.
}
